feat: refactor Hive class to inherit from Entity class
The Hive class has been refactored to inherit from the Entity class. This allows for better code organization and separation of concerns. The Hive class now has an additional parameter in its constructor to accept a world object. The world object is used to access the map and other entities in the world. The print_hive_status method has been renamed to print_world_status and now accepts the world map, number of found food sources, and total food sources as parameters. The add_food_source method has been modified to ensure that the food source is not placed at the same position as the hive. The forage method has been simplified to remove the unnecessary iteration parameter and the map update logic has been moved to the Animator class.